+++ src/predicate/isNotNil.ts
@@ -3,6 +3,7 @@
  *
  * The main use of this function is to be used with TypeScript as a type predicate.
  *
+ * @template T - The type of value.
  * @param {T | null | undefined} x - The value to test if it is not null nor undefined.
  * @returns {x is T} True if the value is not null nor undefined, false otherwise.
  *